data-acquisition/requirements.txt
requests
pika
schedule
python-dotenv

